+++ src/lib/components/forms/form/form-components/text-input/text-input.interface.ts
@@ -6,4 +6,5 @@ export interface TextInputInterface extends TerraFormComponentBaseInterface {
     isReadonly: boolean;
     minLength: number;
     maxLength: number;
+    email: boolean;
 }