+++ src/lib/translations/l10n-terra-components.ts
@@ -95,7 +95,8 @@ const de: {} = {
         }
     },
     terraTextInput: {
-        invalidIban: 'IBAN ist ungültig.'
+        invalidIban: 'IBAN ist ungültig.',
+        invalidEmail: 'E-Mail ist ungültig.'
     },
     terraNodeTree: {
         search: 'Suche'
@@ -243,7 +244,8 @@ const en: {} = {
         }
     },
     terraTextInput: {
-        invalidIban: 'IBAN is not valid.'
+        invalidIban: 'IBAN is not valid.',
+        invalidEmail: 'Email is not valid.'
     },
     terraNodeTree: {
         search: 'Search'